Refactor CNC_Functions; Move Testing Functions to New Testing File
